Move room models in separate package
