+++ app/src/main/kotlin/org/bspb/smartbirds/pro/repository/MonitoringRepository.kt
@@ -2,8 +2,8 @@ package org.bspb.smartbirds.pro.repository
 
 import androidx.lifecycle.LiveData
 import org.bspb.smartbirds.pro.content.Monitoring
-import org.bspb.smartbirds.pro.room.MonitoringModel
-import org.bspb.smartbirds.pro.room.MonitoringWithEntriesCount
+import org.bspb.smartbirds.pro.room.model.MonitoringModel
+import org.bspb.smartbirds.pro.room.model.MonitoringWithEntriesCount
 import org.bspb.smartbirds.pro.room.SmartBirdsRoomDatabase
 
 class MonitoringRepository {
