+++ app/src/main/kotlin/org/bspb/smartbirds/pro/room/dao/FormDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Update
-import org.bspb.smartbirds.pro.room.Form
+import org.bspb.smartbirds.pro.room.model.Form
 
 @Dao
 interface FormDao {
