+++ app/src/main/kotlin/org/bspb/smartbirds/pro/room/dao/ZoneDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
-import org.bspb.smartbirds.pro.room.ZoneModel
+import org.bspb.smartbirds.pro.room.model.ZoneModel
 
 @Dao
 abstract class ZoneDao {
