+++ app/src/main/kotlin/org/bspb/smartbirds/pro/room/model/NomenclatureModel.kt
@@ -1,18 +1,15 @@
-package org.bspb.smartbirds.pro.room
+package org.bspb.smartbirds.pro.room.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity(tableName = "forms")
-data class Form(
+@Entity(tableName = "nomenclatures")
+data class NomenclatureModel(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "_id")
     var id: Long = 0,
-    val code: String,
-    val type: String,
-    val latitude: Double,
-    val longitude: Double,
+    var type: String?,
     @ColumnInfo(typeAffinity = ColumnInfo.BLOB)
-    val data: String
+    var data: String?,
 )
