+++ app/src/main/kotlin/org/bspb/smartbirds/pro/sync/ZonesManager.kt
@@ -9,7 +9,7 @@ import org.androidannotations.annotations.EBean
 import org.androidannotations.annotations.RootContext
 import org.bspb.smartbirds.pro.backend.Backend
 import org.bspb.smartbirds.pro.room.SmartBirdsRoomDatabase
-import org.bspb.smartbirds.pro.room.ZoneModel
+import org.bspb.smartbirds.pro.room.model.ZoneModel
 import org.bspb.smartbirds.pro.tools.Reporting
 import org.bspb.smartbirds.pro.tools.SBGsonParser
 import java.io.IOException
