+++ app/src/main/kotlin/org/bspb/smartbirds/pro/viewmodel/BaseEntryViewModel.kt
@@ -3,7 +3,7 @@ package org.bspb.smartbirds.pro.viewmodel
 import androidx.lifecycle.ViewModel
 import org.bspb.smartbirds.pro.content.MonitoringEntry
 import org.bspb.smartbirds.pro.repository.FormRepository
-import org.bspb.smartbirds.pro.room.Form
+import org.bspb.smartbirds.pro.room.model.Form
 import org.bspb.smartbirds.pro.utils.MonitoringManager
 
 class BaseEntryViewModel : ViewModel() {
