+++ app/src/main/kotlin/org/bspb/smartbirds/pro/viewmodel/MonitoringViewModel.kt
@@ -5,7 +5,7 @@ import org.bspb.smartbirds.pro.backend.dto.Zone
 import org.bspb.smartbirds.pro.content.MonitoringEntry
 import org.bspb.smartbirds.pro.repository.FormRepository
 import org.bspb.smartbirds.pro.repository.ZoneRepository
-import org.bspb.smartbirds.pro.room.Form
+import org.bspb.smartbirds.pro.room.model.Form
 import org.bspb.smartbirds.pro.utils.MonitoringManager
 
 class MonitoringViewModel : ViewModel() {
